+++ packages/protocol/contracts/L1/libs/LibUtils.sol
@@ -41,7 +41,7 @@ library LibUtils {
     {
         (TaikoData.Block storage blk, uint64 slot) = getBlock(_state, _config, _blockId);
 
-        uint32 tid = getTransitionId(_state, blk, slot, _parentHash);
+        uint24 tid = getTransitionId(_state, blk, slot, _parentHash);
         if (tid == 0) revert L1_TRANSITION_NOT_FOUND();
 
         return _state.transitions[slot][tid];
@@ -58,7 +58,7 @@ library LibUtils {
         TaikoData.State storage _state,
         TaikoData.Config memory _config,
         uint64 _blockId,
-        uint32 _tid
+        uint24 _tid
     )
         internal
         view
@@ -98,7 +98,7 @@ library LibUtils {
     )
         internal
         view
-        returns (uint32 tid_)
+        returns (uint24 tid_)
     {
         if (_state.transitions[_slot][1].key == _parentHash) {
             tid_ = 1;
